Refactor ItemList
