+++ client/src/sections/Fridge/ItemList/AddItemModal/index.ts
@@ -1,0 +1,3 @@
+import AddItemModal from './AddItemModal';
+
+export default AddItemModal;
