+++ client/src/sections/Fridge/ItemList/ItemList.css
@@ -5,7 +5,7 @@
     align-items: center;
     justify-content: flex-start;
     background-color: #5C6AC4;
-    padding: 3rem;
+    padding: 1rem;
 }
 
 .item {
