+++ client/src/sections/Fridge/ItemList/components.ts
@@ -1,1 +1,2 @@
 export {default as Item} from './Item';
+export {default as AddItemModal} from './AddItemModal';
